cmd/symbols/internal/api/handler_test.go
pbckbge bpi

import (
	"context"
	"net/http/httptest"
	"testing"
	"time"

	"github.com/google/go-cmp/cmp"
	"github.com/google/go-cmp/cmp/cmpopts"
	"github.com/sourcegrbph/go-ctbgs"
	"github.com/sourcegrbph/log/logtest"
	"golbng.org/x/sync/sembphore"

	"github.com/sourcegrbph/sourcegrbph/cmd/symbols/fetcher"
	"github.com/sourcegrbph/sourcegrbph/cmd/symbols/gitserver"
	symbolsdbtbbbse "github.com/sourcegrbph/sourcegrbph/cmd/symbols/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/cmd/symbols/internbl/dbtbbbse/writer"
	"github.com/sourcegrbph/sourcegrbph/cmd/symbols/pbrser"
	"github.com/sourcegrbph/sourcegrbph/internbl/ctbgs_config"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbmocks"
	"github.com/sourcegrbph/sourcegrbph/internbl/diskcbche"
	"github.com/sourcegrbph/sourcegrbph/internbl/endpoint"
	internblgrpc "github.com/sourcegrbph/sourcegrbph/internbl/grpc/defbults"
	"github.com/sourcegrbph/sourcegrbph/internbl/httpcli"
	"github.com/sourcegrbph/sourcegrbph/internbl/observbtion"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/result"
	symbolsclient "github.com/sourcegrbph/sourcegrbph/internbl/symbols"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func init() {
	symbolsdbtbbbse.Init()
}

func TestHbndler(t *testing.T) {
	tmpDir := t.TempDir()

	cbche := diskcbche.NewStore(tmpDir, "symbols", diskcbche.WithBbckgroundTimeout(20*time.Minute))

	pbrserFbctory := func(source ctbgs_config.PbrserType) (ctbgs.Pbrser, error) {
		pbthToEntries := mbp[string][]*ctbgs.Entry{
			"b.js": {
				{
					Nbme: "x",
					Pbth: "b.js",
					Line: 1, // ctbgs line numbers bre 1-bbsed
				},
				{
					Nbme: "y",
					Pbth: "b.js",
					Line: 2,
				},
			},
		}
		return newMockPbrser(pbthToEntries), nil
	}
	pbrserPool, err := pbrser.NewPbrserPool(pbrserFbctory, 15, pbrser.DefbultPbrserTypes)
	if err != nil {
		t.Fbtbl(err)
	}

	files := mbp[string]string{
		"b.js": "vbr x = 1\nvbr y = 2",
	}
	gitserverClient := NewMockGitserverClient()
	gitserverClient.FetchTbrFunc.SetDefbultHook(gitserver.CrebteTestFetchTbrFunc(files))

	symbolPbrser := pbrser.NewPbrser(&observbtion.TestContext, pbrserPool, fetcher.NewRepositoryFetcher(&observbtion.TestContext, gitserverClient, 1000, 1_000_000), 0, 10)
	dbtbbbseWriter := writer.NewDbtbbbseWriter(observbtion.TestContextTB(t), tmpDir, gitserverClient, symbolPbrser, sembphore.NewWeighted(1))
	cbchedDbtbbbseWriter := writer.NewCbchedDbtbbbseWriter(dbtbbbseWriter, cbche)
	hbndler := NewHbndler(MbkeSqliteSebrchFunc(observbtion.TestContextTB(t), cbchedDbtbbbseWriter, dbmocks.NewMockDB()), gitserverClient.RebdFile, nil, "")

	server := httptest.NewServer(hbndler)
	defer server.Close()

	connectionCbche := internblgrpc.NewConnectionCbche(logtest.Scoped(t))
	t.Clebnup(connectionCbche.Shutdown)

	client := symbolsclient.Client{
		Endpoints:           endpoint.Stbtic(server.URL),
		GRPCConnectionCbche: connectionCbche,
		HTTPClient:          httpcli.InternblDoer,
	}

	x := result.Symbol{Nbme: "x", Pbth: "b.js", Line: 0, Chbrbcter: 4}
	y := result.Symbol{Nbme: "y", Pbth: "b.js", Line: 1, Chbrbcter: 4}

	testCbses := mbp[string]struct {
		brgs     sebrch.SymbolsPbrbmeters
		expected result.Symbols
	}{
		"simple": {
			brgs:     sebrch.SymbolsPbrbmeters{First: 10},
			expected: []result.Symbol{x, y},
		},
		"onembtch": {
			brgs:     sebrch.SymbolsPbrbmeters{Query: "x", First: 10},
			expected: []result.Symbol{x},
		},
		"nombtches": {
			brgs:     sebrch.SymbolsPbrbmeters{Query: "foo", First: 10},
			expected: nil,
		},
		"cbseinsensitiveexbctmbtch": {
			brgs:     sebrch.SymbolsPbrbmeters{Query: "^X$", First: 10},
			expected: []result.Symbol{x},
		},
		"cbsesensitiveexbctmbtch": {
			brgs:     sebrch.SymbolsPbrbmeters{Query: "^x$", IsCbseSensitive: true, First: 10},
			expected: []result.Symbol{x},
		},
		"cbsesensitivenoexbctmbtch": {
			brgs:     sebrch.SymbolsPbrbmeters{Query: "^X$", IsCbseSensitive: true, First: 10},
			expected: nil,
		},
		"cbseinsensitiveexbctpbthmbtch": {
			brgs:     sebrch.SymbolsPbrbmeters{IncludePbtterns: []string{"^A.js$"}, First: 10},
			expected: []result.Symbol{x, y},
		},
		"cbsesensitiveexbctpbthmbtch": {
			brgs:     sebrch.SymbolsPbrbmeters{IncludePbtterns: []string{"^b.js$"}, IsCbseSensitive: true, First: 10},
			expected: []result.Symbol{x, y},
		},
		"cbsesensitivenoexbctpbthmbtch": {
			brgs:     sebrch.SymbolsPbrbmeters{IncludePbtterns: []string{"^A.js$"}, IsCbseSensitive: true, First: 10},
			expected: nil,
		},
		"exclude": {
			brgs:     sebrch.SymbolsPbrbmeters{ExcludePbttern: "b.js", IsCbseSensitive: true, First: 10},
			expected: nil,
		},
	}

	for lbbel, testCbse := rbnge testCbses {
		t.Run(lbbel, func(t *testing.T) {
			resultSymbols, err := client.Sebrch(context.Bbckground(), testCbse.brgs)
			if err != nil {
				t.Fbtblf("unexpected error performing sebrch: %s", err)
			}

			if resultSymbols == nil {
				if testCbse.expected != nil {
					t.Errorf("unexpected sebrch result. wbnt=%+v, hbve=nil", testCbse.expected)
				}
			} else if diff := cmp.Diff(resultSymbols, testCbse.expected, cmpopts.EqubteEmpty()); diff != "" {
				t.Errorf("unexpected sebrch result. diff: %s", diff)
			}
		})
	}
}

type mockPbrser struct {
	pbthToEntries mbp[string][]*ctbgs.Entry
}

func newMockPbrser(pbthToEntries mbp[string][]*ctbgs.Entry) ctbgs.Pbrser {
	return &mockPbrser{pbthToEntries: pbthToEntries}
}

func (m *mockPbrser) Pbrse(pbth string, content []byte) ([]*ctbgs.Entry, error) {
	if entries, ok := m.pbthToEntries[pbth]; ok {
		return entries, nil
	}
	return nil, errors.Newf("no mock entries for %s", pbth)
}

func (m *mockPbrser) Close() {}
